refactor: routes, layout 수정
