--- src/pages/place/culture/index.tsx
@@ -1,3 +1,0 @@
-export default function Culture() {
-  return <div>Culture</div>;
-}
